refactor(tests): Move content type, CSRF token, and secure headers tests from `ServerRequestAdapterTest` to new `HeadersPsr7Test` class for better organization and clarity.
